+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reordered API structure by moving session API above the AuthMiddleware.
 - Modified user authentication flow: 'Last Login' information is now stored in the browser's localStorage.
 - Completly refactored the authentication mechanism in the app by replacing sessions with JWT Tokens
+- Countdown expiry displayed in red
 
 ## [v15.20.0] - 2024-01-28